translation_project/requirements.txt
absl-py==0.12.0
aiohttp==3.5.4
APScheduler==3.2.0
astroid==2.3.3
astunparse==1.6.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==18.2.0
beautifulsoup4==4.7.1
cached-property==1.5.2
cachetools==4.2.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.3
cssselect==1.1.0
cssutils==1.0.2
cycler==0.10.0
dnspython==2.1.0
docx2python==1.27.1
docxcompose==1.3.1
et-xmlfile==1.1.0
eventlet==0.32.0
fire==0.4.0
Flask==0.12
flatbuffers==1.12
fonttools==4.29.0
gast==0.4.0
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googletrans==2.4.0
greenlet==1.1.1
grpcio==1.34.1
h5py==3.1.0
html2text==2020.1.16
html5lib==1.0.1
htmlparsing==0.1.5
idna==2.7
imbox==0.9.8
importlib-metadata==4.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10
JPype1==1.2.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.0.1
langid==1.1.6
latex2mathml==3.0.0
lazy-object-proxy==1.4.3
lxml==4.3.0
Markdown==3.3.4
MarkupSafe==1.1.0
mathml2omml==0.0.2
matplotlib==3.0.2
mccabe==0.6.1
module==0.0.4
multidict==4.5.2
multilingual-pdf2text==1.1.0
mypy==0.761
mypy-extensions==0.4.3
nltk==3.4.5
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.5.62
openpyxl==3.0.7
opt-einsum==3.3.0
pandas==1.1.5
parse==1.19.0
pdf2docx==0.5.2
pdf2image==1.16.0
pdfminer==20191125
pdfminer.six==20200104
pdfminer3k==1.3.4
pdfplumber==0.5.16
Pillow==7.0.0
playsound==1.2.2
ply==3.11
premailer==3.7.0
protobuf==3.17.3
psutil==5.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.4
pydantic==1.9.0
PyExecJS==1.5.1
pygal==2.4.0
pylint==2.4.4
PyMuPDF==1.19.4
pynput==1.6.6
pyparsing==2.3.1
pyPdf==1.13
PyPDF2==1.26.0
pyperclip==1.7.0
pypiwin32==223
pytesseract==0.3.8
python-dateutil==2.7.5
python-docx==0.8.10
pytz==2018.9
pywin32==301
redis==3.2.0
requests==2.25.1
requests-futures==0.9.9
requests-oauthlib==1.3.0
rsa==4.7.2
selenium==3.141.0
six==1.15.0
some-package==0.1
sortedcontainers==2.1.0
soupsieve==1.7.3
tencentcloud-sdk-python==3.0.125
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
Transcrypt==3.7.16
typed-ast==1.4.1
typing_extensions==4.0.1
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.24.1
Wand==0.5.8
webencodings==0.5.1
Werkzeug==0.11.15
wrapt==1.12.1
yagmail==0.14.245
yarl==1.3.0
zipp==3.4.1
zmail==0.2.8
